mcp-server/requirements.txt
fastapi>=0.68.0
uvicorn[standard]>=0.15.0
python-jose>=3.3.0
python-multipart>=0.0.5
redis>=4.0.0
python-dotenv>=0.19.0
pydantic>=1.8.0
pydantic-settings>=2.0.0
asyncpg>=0.27.0
aioredis>=2.0.0
tenacity>=8.0.0
loguru>=0.5.0
elasticsearch[async]==8.9.0
PyJWT>=2.8.0
SQLAlchemy>=1.4
aiohttp>=3.9.0
fastmcp==2.7.0
mcp>=1.9.3
rich>=13.9.4
httpx>=0.28.1